@@ -1,16 +1,11 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import "./globals.css";
-import { Asap } from "next/font/google";
+import { ThemeProvider } from "./context/ThemeProvider";
 import { AuthProvider } from "./context/AuthContext";
 import { GoogleOAuthProvider } from "@react-oauth/google";
-
-const inter = Inter({
-  subsets: ["latin"],
-  variable: "--font-inter",
-});
-
-const asap = Asap({ subsets: ["latin"], weight: ["400", "700"] });
+import { NextIntlClientProvider } from 'next-intl';
+import { getLocale } from 'next-intl/server';
+import { defaultLocale } from '../../i18n';
 
 export const metadata: Metadata = {
   title: "Pet Shop",
@@ -19,13 +14,25 @@ export const metadata: Metadata = {
 
 const GOOGLE_CLIENT_ID = process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID || "";
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
+  // Lấy locale từ middleware
+  const locale = await getLocale() || defaultLocale;
+  let messages;
+  try {
+    // Tải file bản dịch cho locale hiện tại
+    messages = (await import(`../locales/${locale}/common.json`)).default;
+  } catch (error) {
+    console.error(`Failed to load messages for locale ${locale}:`, error);
+    // Fallback về locale mặc định (vi)
+    messages = (await import(`../locales/vi/common.json`)).default;
+  }
+
   return (
-    <html lang="en">
+    <html lang={locale} suppressHydrationWarning>
       <head>
         <link
           rel="icon"
@@ -33,14 +40,20 @@ export default function RootLayout({
           type="image/png"
         />
       </head>
-      <body className={`${inter.className} ${asap.className} antialiased`}>
-        {GOOGLE_CLIENT_ID ? (
-          <GoogleOAuthProvider clientId={GOOGLE_CLIENT_ID}>
-            <AuthProvider>{children}</AuthProvider>
-          </GoogleOAuthProvider>
-        ) : (
-          <p>Error: Missing Google Client ID</p>
-        )}
+      <body>
+        <NextIntlClientProvider locale={locale} messages={messages}>
+          <ThemeProvider>
+            {GOOGLE_CLIENT_ID ? (
+              <GoogleOAuthProvider clientId={GOOGLE_CLIENT_ID}>
+                <AuthProvider>
+                  {children}
+                </AuthProvider>
+              </GoogleOAuthProvider>
+            ) : (
+              <p>Error: Missing Google Client ID</p>
+            )}
+          </ThemeProvider>
+        </NextIntlClientProvider>
       </body>
     </html>
   );
