@@ -1,5 +1,6 @@
 import React from "react";
 import Image from "next/image";
+import Link from "next/link";
 
 const Footer = () => {
   return (
@@ -31,44 +32,44 @@ const Footer = () => {
             <h3 className="text-lg font-bold mb-4">Shop</h3>
             <ul className="space-y-2">
               <li>
-                <a
+                <Link
                   href="/collections/thuc-an-cho-cho"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Dành Cho Chó
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/collections/mua-sam-cho-meo"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Dành Cho Mèo
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/pages/brands-thuong-hieu-thu-cung"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Thương Hiệu
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/blogs/cham-soc-thu-cung"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Blogs
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/pages/brands-thuong-hieu-thu-cung"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Bộ Sưu Tập
-                </a>
+                </Link>
               </li>
             </ul>
           </div>
@@ -77,36 +78,36 @@ const Footer = () => {
             <h3 className="text-lg font-bold mb-4">Paddy Pet Shop</h3>
             <ul className="space-y-2">
               <li>
-                <a
+                <Link
                   href="/pages/paddy-pet-shop"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Giới Thiệu
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/pages/uu-dai-tich-luy-thanh-vien-paddier"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Thành Viên Paddier
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/policies/terms-of-service"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Điều Khoản Sử Dụng
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/pages/tuyen-dung"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Tuyển Dụng
-                </a>
+                </Link>
               </li>
             </ul>
           </div>
@@ -115,44 +116,44 @@ const Footer = () => {
             <h3 className="text-lg font-bold mb-4">Hỗ Trợ Khách Hàng</h3>
             <ul className="space-y-2">
               <li>
-                <a
+                <Link
                   href="/pages/chinh-sach-doi-tra-hang"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Chính Sách Đổi Trả Hàng
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/policies/shipping-policy"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Phương Thức Vận Chuyển
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/policies/privacy-policy"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Chính Sách Bảo Mật
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/pages/huong-dan-thanh-toan"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Phương Thức Thanh Toán
-                </a>
+                </Link>
               </li>
               <li>
-                <a
+                <Link
                   href="/policies/refund-policy"
                   className="text-gray-700 hover:text-blue-600"
                 >
                   Chính Sách Hoàn Tiền
-                </a>
+                </Link>
               </li>
             </ul>
           </div>
@@ -170,19 +171,19 @@ const Footer = () => {
               </p>
               <p className="mt-4">
                 <span className="font-bold">Hotline:</span>{" "}
-                <a href="tel:0867677891" className="text-blue-600">
+                <Link href="tel:0867677891" className="text-blue-600">
                   0867677891
-                </a>
+                </Link>
               </p>
               <p className="mt-2">
                 <span className="font-bold">Email:</span>{" "}
-                <a href="mailto:contact@paddy.vn" className="text-blue-600">
+                <Link href="mailto:contact@paddy.vn" className="text-blue-600">
                   contact@paddy.vn
-                </a>
+                </Link>
               </p>
             </div>
             <div className="mt-6 flex space-x-2 ">
-              <a
+              <Link
                 href="https://www.facebook.com/PaddyPetShop"
                 className="text-gray-700 hover:text-blue-600 border border-transparent rounded-full p-2 hover:border-blue-600 transition-all duration-300"
               >
@@ -194,8 +195,8 @@ const Footer = () => {
                 >
                   <path d="M24 12.073c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.99 4.388 10.954 10.125 11.854v-8.385H7.078v-3.47h3.047V9.43c0-3.007 1.792-4.669 4.533-4.669 1.312 0 2.686.234 2.686.234v2.953h-1.508c-1.491 0-1.956.925-1.956 1.874v2.25h3.328l-.532 3.47h-2.796v8.385C19.612 23.027 24 18.062 24 12.073z" />
                 </svg>
-              </a>
-              <a
+              </Link>
+              <Link
                 href="https://www.instagram.com/paddypetshop/"
                 className="text-gray-700 hover:text-blue-600 border border-transparent rounded-full p-2 hover:border-blue-600 transition-all duration-300"
               >
@@ -207,8 +208,8 @@ const Footer = () => {
                 >
                   <path d="M12 2.163c3.204 0 3.584.012 4.85.07 3.252.148 4.771 1.691 4.919 4.919.058 1.265.069 1.645.069 4.849 0 3.205-.012 3.584-.069 4.849-.149 3.225-1.664 4.771-4.919 4.919-1.266.058-1.644.07-4.85.07-3.204 0-3.584-.012-4.849-.07-3.26-.149-4.771-1.699-4.919-4.92-.058-1.265-.07-1.644-.07-4.849 0-3.204.013-3.583.07-4.849.149-3.227 1.664-4.771 4.919-4.919 1.266-.057 1.645-.069 4.849-.069zm0-2.163c-3.259 0-3.667.014-4.947.072-4.358.2-6.78 2.618-6.98 6.98-.059 1.281-.073 1.689-.073 4.948 0 3.259.014 3.668.072 4.948.2 4.358 2.618 6.78 6.98 6.98 1.281.058 1.689.072 4.948.072 3.259 0 3.668-.014 4.948-.072 4.354-.2 6.782-2.618 6.979-6.98.059-1.28.073-1.689.073-4.948 0-3.259-.014-3.667-.072-4.947-.196-4.354-2.617-6.78-6.979-6.98-1.281-.059-1.69-.073-4.949-.073zm0 5.838c-3.403 0-6.162 2.759-6.162 6.162s2.759 6.163 6.162 6.163 6.162-2.759 6.162-6.163c0-3.403-2.759-6.162-6.162-6.162zm0 10.162c-2.209 0-4-1.79-4-4 0-2.209 1.791-4 4-4s4 1.791 4 4c0 2.21-1.791 4-4 4zm6.406-11.845c-.796 0-1.441.645-1.441 1.44s.645 1.44 1.441 1.44c.795 0 1.439-.645 1.439-1.44s-.644-1.44-1.439-1.44z" />
                 </svg>
-              </a>
-              <a
+              </Link>
+              <Link
                 href="https://www.tiktok.com/@paddypetshop"
                 className="text-gray-700 hover:text-blue-600 border border-transparent rounded-full p-2 hover:border-blue-600 transition-all duration-300"
               >
@@ -220,13 +221,13 @@ const Footer = () => {
                 >
                   <path d="M19.59 6.69a4.83 4.83 0 0 1-3.77-4.25V2h-3.45v13.67a2.89 2.89 0 0 1-5.2 1.74 2.89 2.89 0 0 1 2.31-4.64 2.93 2.93 0 0 1 .88.13V9.4a6.84 6.84 0 0 0-1-.05A6.33 6.33 0 0 0 5 20.1a6.34 6.34 0 0 0 10.86-4.43v-7a8.16 8.16 0 0 0 4.77 1.52v-3.4a4.85 4.85 0 0 1-1-.1z" />
                 </svg>
-              </a>
+              </Link>
             </div>
           </div>
         </div>
       </div>
 
-      <div className="bg-white py-6 flex justify-between items-center">
+      <div className="py-6 flex justify-between items-center">
         <div className="mt-6 ml-20 px-4">
           <p className="text-gray-400">@2024 Paddy VN. All Rights Reserved.</p>
         </div>
